docker_proyect/requirements.txt
numpy==1.26.4
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
sahi==0.11.16
ultralytics==8.2.32
